+++ cts/lab/CTStests.py
@@ -26,45 +26,10 @@ from stat import *
 from pacemaker import BuildOptions
 from pacemaker._cts.CTS import NodeStatus
 from pacemaker._cts.audits import AuditResource
-from pacemaker._cts.tests import CTSTest, RemoteDriver, SimulStartLite, SimulStopLite, StartTest, StopTest
+from pacemaker._cts.tests import *
 from pacemaker._cts.timer import Timer
 
 AllTestClasses = [ ]
-
-
-class FlipTest(CTSTest):
-    '''If it's running, stop it.  If it's stopped start it.
-       Overthrow the status quo...
-    '''
-    def __init__(self, cm):
-        CTSTest.__init__(self,cm)
-        self.name = "Flip"
-        self._start = StartTest(cm)
-        self._stop = StopTest(cm)
-
-    def __call__(self, node):
-        '''Perform the 'Flip' test. '''
-        self.incr("calls")
-        if self._cm.ShouldBeStatus[node] == "up":
-            self.incr("stopped")
-            ret = self._stop(node)
-            type = "up->down"
-            # Give the cluster time to recognize it's gone...
-            time.sleep(self._env["StableTime"])
-        elif self._cm.ShouldBeStatus[node] == "down":
-            self.incr("started")
-            ret = self._start(node)
-            type = "down->up"
-        else:
-            return self.skipped()
-
-        self.incr(type)
-        if ret:
-            return self.success()
-        else:
-            return self.failure("%s failure" % type)
-
-#        Register FlipTest as a good test to run
 AllTestClasses.append(FlipTest)
